Do not provide a feature in features/support/env.el
This is not a library intended to be loaded with `require', so it
should not advertise itself as such by `provide'ing `env'.  Doing that
would cause conflicts with other packages that use `ecukes'.  This
file is automatically `load'ed by `ecukes-load-support', defined in
"ecukes-load.el".  Other test related files of a project don't have to
load it explicitly.